pythonImageAnalysis/requirements.txt
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.4.0
psutil==5.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
